[83] Out-date templates
Main point of this PR is to fix and improve the out-date examples
of Odahu APIs.
But accidentally discovered, that the get_odahuflow_template_names
function has been broken after the recent renaming. So the CLI command
and Jupyterlab integration has been broken as well.
Covered this problem by tests to prevent relapse in the future.
In addition, the file parameter of generate template command was
removed because it can be replaced by redirection of stdout to a file.