[feat] Various Odahuflowctl improvement
Stopped to show an error traceback to user. For now we produce
small message when an exception was occurred.

Any logs, including error level, doesn't output until a user
provides the --verbose flag.

Enabled a shell completion for CLI.

Added aliases for the commands in their documentation.

Removed obsolete the sandbox command.