Fix mistaken re-creation of Knative Revisions caused by change-watching logic (#530)
Previously, it was based on k8s resourceVersion of an MD. resourceVersion of latest applied MD was stored in annotations of a Knative Service. If an MD gets new resourceVersion, ODAHU operator updated KService with appropriate spec.
The problem is that resourceVersion updates even for changes in .Status object. So there were a lot of unnecessary revisions spawned because of status updates.

Now It honestly stores last applied MD spec in annotations and compares it with an actual MD state. It differs, then KService must be updated.